# packages/vische-server/swagger-spec/swagger-spec.yaml
openapi: 3.0.0
info:
  title: VischeAPI
  description: ''
  version: 1.0.0
  contact: {}
...
servers: []
components:
  schemas:
    PostChannelListRequest:
      type: object
      properties:
        active_only:
          type: boolean
          default: true
        channel_id:
          default: []
          type: array
          items:
            type: string
        group_id:
          default: []
          type: array
          items:
            type: string
        sub_group_id:
          default: []
          type: array
          items:
            type: string
      required:
        - active_only
        - channel_id
        - group_id
        - sub_group_id
    Channel:
      type: object
      properties:
        channel_id:
          type: string
        channel_title:
          type: string
        enable:
          type: boolean
        name:
          type: string
        published_date:
          type: string
        icon_url:
          type: string
        country:
          type: string
        group_id:
          type: string
        sub_group_id:
          type: string
        twitterId:
          type: string
        stream_hash_tag:
          type: array
          items:
            type: string
        color:
          type: string
      required:
        - channel_id
        - channel_title
        - enable
        - name
        - published_date
        - icon_url
        - country
        - group_id
        - sub_group_id
        - twitterId
        - stream_hash_tag
        - color
    PostChannelListResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/Channel'
      required:
        - status
        - message
        - data
...
paths:
...
  /api/channel/list:
    post:
      operationId: ChannelsController_getChannels
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostChannelListRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostChannelListResponse'
...